# openapi.yaml
openapi: 3.0.3
info:
  title: Fleet Management System API
  description: |
    Comprehensive API for managing school transportation operations including 
    bus fleets, driver assignments, student ridership, and vehicle maintenance.
  version: 1.0.0
  contact:
    name: Fleet Management Support
    email: support@fleetmanagement.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.fleetmanagement.com
    description: Production server
...
security:
  - sessionAuth: []
  - csrfToken: []
paths:
...
  /add-bus:
    post:
      summary: Add a new bus to the fleet
      tags:
        - Fleet
      security:
        - sessionAuth: []
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddBusRequest'
      responses:
        '200':
          description: Bus added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  bus_id:
                    type: string
...
  /add-student:
    post:
      summary: Add a new student
      tags:
        - Students
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AddStudentRequest'
      responses:
        '303':
          description: Student added successfully
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session_id
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  schemas:
...
    AddBusRequest:
      type: object
      required:
        - bus_id
        - model
        - capacity
        - status
        - csrf_token
      properties:
        bus_id:
          type: string
          pattern: '^[A-Z0-9]+$'
        model:
          type: string
        capacity:
          type: integer
          minimum: 1
          maximum: 100
        status:
          type: string
          enum: [active, maintenance, out of service]
        oil_status:
          type: string
          enum: [good, due soon, overdue]
        tire_status:
          type: string
          enum: [good, fair, poor]
        maintenance_notes:
          type: string
        csrf_token:
          type: string
    AddStudentRequest:
      type: object
      required:
        - student_id
        - name
        - phone_number
        - guardian
        - pickup_time
        - dropoff_time
        - csrf_token
      properties:
        student_id:
          type: string
        name:
          type: string
        phone_number:
          type: string
        alt_phone_number:
          type: string
        guardian:
          type: string
        pickup_time:
          type: string
          pattern: '^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$'
        dropoff_time:
          type: string
          pattern: '^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$'
        position_number:
          type: integer
        route_id:
          type: string
        csrf_token:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
...
  responses:
...
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: VALIDATION_ERROR
              message: Invalid input data
...